requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bottle==0.13.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
Django==5.2
djangorestframework==3.16.0
dnspython==2.7.0
docutils==0.21.2
fastapi==0.115.12
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
instagrapi==2.1.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
libtool==1.1.0
Markdown==3.7
Markdown-Editor==1.0.7
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
nes_py==8.2.1
nh3==0.2.21
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyglet==1.5.21
Pygments==2.19.1
pymdown-extensions==7.1
pymongo==4.12.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
python-telegram-bot==22.0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
setuptools==78.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
tqdm==4.67.1
twine==6.1.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.46.1
